backend/requirements.txt
# Django
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0

# Database
psycopg2-binary>=2.9,<3.0

# Redis & Celery
redis>=5.0,<6.0
celery>=5.3,<6.0

# Environment variables
python-dotenv>=1.0,<2.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Development
ipython>=8.0
